+++ healthper/src/main/java/umc/healthper/service/post/PostLikeService.java
@@ -1,17 +1,17 @@
-package umc.healthper.service;
+package umc.healthper.service.post;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import umc.healthper.domain.member.Member;
 import umc.healthper.domain.post.Post;
-import umc.healthper.domain.like.PostLike;
+import umc.healthper.domain.post.PostLike;
 import umc.healthper.exception.member.MemberNotFoundByIdException;
 import umc.healthper.exception.post.PostNotFoundException;
 import umc.healthper.exception.postlike.AlreadyPostLikeException;
 import umc.healthper.exception.postlike.PostLikeNotFoundException;
 import umc.healthper.repository.MemberRepository;
-import umc.healthper.repository.like.PostLikeRepository;
+import umc.healthper.repository.post.PostLikeRepository;
 import umc.healthper.repository.post.PostRepository;
 
 @Service
